[PATCH] Accept an ACK after a NAK
Cherry-picked from
https://chromium.googlesource.com/chromiumos/overlays/chromiumos-overlay/+/
master/net-misc/dhcpcd/files/patches/dhcpcd-6.8.2-Accept-an-ACK-after-a-
NAK.patch.

When a NAK is received, wait for a response interval before
acting on it.  In doing so, this allows for networks that
may have duelling DHCP servers which both consider themselves
authoritative on a network.  Since other DHCP clients also
act in a similar manner, issues of this sort end up undetected
and network administrators thus consider this situation nominal.

BUG: 22956197
Change-Id: I8cceb03f8296abb4ee6bbdc87c222ae9234b3f5f

Reviewed-on: https://chromium-review.googlesource.com/208207