[PATCH] Merge in DHCP options from the orignal offer
Cherry-picked from
https://chromium.googlesource.com/chromiumos/overlays/chromiumos-overlay/+/
master/net-misc/dhcpcd/files/patches/dhcpcd-6.8.2-Merge-in-DHCP-options-
from-the-original-offer.patch.

We've found that some APs respond to DHCP REQUEST messages with a
subset of the DHCP options that were present in the original DHCP
negotiation.  Copy such options out of the stored lease and carry
them forward whenever a lease renewal succeeds.

BUG: 22956197
Change-Id: I9fa006535a67cdb3e25f7d02a8a81d051033922e

Reviewed-on: https://chromium-review.googlesource.com/195270