dhcpcd-6.8.2: assign interface index in get_interface
In if_discover, the interface struct's index member
(ifp->index) is assigned from the sll_ifindex member of
of ifa->if_addr (see if.c:490). However, this member is
not set in the get_interface function of ifaddrs.c. This
unitialized interface index member is later used to bind
a socket and send messages on it, which leads to errors.

Assign the interface index of selected interface in
get_interface to fix this.

BUG: 22956197
Change-Id: I41da8fc7abdf3f5fc05c1dc0446d2a1d9dc7557d